fixed auto and teleop

+++ SkyStone-master/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/blueColorAutoSimple.java
@@ -180,7 +180,7 @@ public class blueColorAutoSimple extends LinearOpMode {
                 telemetry.addData("Hue", hsvValues[0]);
                 telemetry.update();
                 //if (hsvValues[0] > 150 ){ // Checks if it is red or blue
-                if(sensorDistance.getDistance(DistanceUnit.CM) < 10)  {
+                if((int)(sensorDistance.getDistance(DistanceUnit.CM)) < 10)  {
                     step++;
                 }
             }
